% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{crossv_mc}
\alias{crossv_kfold}
\alias{crossv_mc}
\title{Generate cross-validated test-training pairs}
\usage{
crossv_mc(data, n, test = 0.2, id = ".id")
crossv_kfold(data, k = 5, id = ".id")
}
\description{
Generate cross-validated test-training pairs
}
\keyword{internal}

