% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelearner.R
\name{is.pipelearner}
\alias{is.pipelearner}
\title{Reports whether x is a pipelearner object}
\usage{
is.pipelearner(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a pipelearner object
}
\keyword{internal}
