% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{learn}
\alias{learn}
\title{Learn from data}
\usage{
learn(pl)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
\code{learn} takes a pipelearner object and conducts the steps needed to
learn/fit all of the specified models on data.
}

