% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_cvpairs.R
\name{learn_cvpairs}
\alias{learn_cvpairs}
\title{Cross-validation pairs}
\usage{
learn_cvpairs(pl, .f = NULL, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{.f}{a function which takes a \code{data} argument}

\item{...}{additional parameters to be passed to \code{.f}}
}
\description{
Add cross validation pairs to a pipelearner object. By default will create
a random partition of the data. Can use any function which returns a data
data frame with three columns: \code{train}, \code{test}, and \code{.id}.
\code{test} and \code{train} are list-columns with each entry of type
\code{resample}. \code{.id} is an atomic vector. If no function is specified,
will create a random partition with proportion 0.2 reserved for testing data.
See \code{\link[modelr]{crossv_mc}} for more details.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}
\examples{
pl <- pipelearner(mtcars)
learn_cvpairs(pl, crossv_kfold, k = 4)
}
