% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{learn_model}
\alias{learn_model}
\title{Machine learning models}
\usage{
learn_model(pl, model, formulas, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{model}{Function call to the learning model}

\item{formulas}{List of objects of class "formula" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{...}{Additional arguments to be passed to the model function as
hyperparameters}
}
\description{
Add a learning model to a pipelearner object.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}

