% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{learn_models}
\alias{learn_models}
\title{Machine learning models}
\usage{
learn_models(pl, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{...}{Function calls of the learning models to use}
}
\description{
Add a learning model to a pipelearner object.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}

