% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_from_resample}
\alias{p_from_resample}
\title{Proportion of resample object}
\usage{
p_from_resample(resample, p)
}
\arguments{
\item{resample}{\code{\link[modelr]{resample}} object}

\item{p}{Non-zero proportion.}
}
\description{
Create a new resample object from the rounded proportion of indices starting
at the beginning of a resample object. For example, create a resample object
pointing to the the first .3 (30\%) or .5 (50\%) of rows pointed to by the
original resample object.
}
