% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelearner.R
\name{pipelearner}
\alias{pipelearner}
\title{Create a new pipe learner}
\usage{
pipelearner(data = NULL)
}
\arguments{
\item{data}{Data frame containing all variables needed by learning algorithm
formulae.}
}
\description{
\code{pipelearner} initializes a new pipelearner object. It is used to
specify the input data frame to subsequent machine learning algorithms.
}

