% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelearner.R
\name{pipelearner_params}
\alias{pipelearner_params}
\title{Parameters shared by pipelearner methods.}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
Parameters shared by pipelearner methods.
}
