% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.pipelearner}
\alias{predict.pipelearner}
\title{Predictions for fitted learning models in a pipelearner object}
\usage{
\method{predict}{pipelearner}(pl)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
Produces true and predicted values for all training and test sets, for all
models, in a pipelearner object.
}
