% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{recover_fits}
\alias{recover_fits}
\title{Recover model and cross-validation information for fits}
\usage{
recover_fits(pl)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
\code{recover_fits} will use the models and cv_pairs `.id` references from a
fits tibble in a pipelearner object to recover relevant information.
}

