% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.pipelearner}
\alias{summary.pipelearner}
\title{Summarize a pipelearner object}
\usage{
\method{summary}{pipelearner}(pl)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
Summarize all fitted models in a pipelearner object
}

