% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn}
\alias{learn}
\title{Learn from data}
\usage{
learn(pl)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}
}
\description{
\code{learn} conducts the steps needed to learn/fit all of the specified
models on all cross-validation training data sets, in all variations of
training sample size, in a pipelearner object, and return the results as a
summarized tibble.
}

