% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curves.R
\name{learn_curves}
\alias{learn_curves}
\title{Learning curves}
\usage{
learn_curves(pl, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{...}{Numbers that are greater than zero and less than or equal to one
representing the proportions of training data to learn on. Will default to
1 if no values are provided.}
}
\description{
Add learning curves (instructions to train on different proportions of each
training set) to a pipelearner object.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}

