% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_curves.R
\name{learn_curves}
\alias{learn_curves}
\title{Learning curves}
\usage{
learn_curves(pl, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{...}{Non-zero proportions of training data to learn on. Will default to
1 if no values are provided.}
}
\description{
Add learning curves (instructions to train on different proportions of each
training set) to a pipelearner object.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.

These proportions are applied to cross-validation training sets using
\code{\link{p_from_resample}}, meaning they will always slice the proportion
starting the beginning. For example, a proportion of .5 will take the first
50\% of rows from the training data (rounded to the nearest whole number).
There is no point taking random subsamples of size p from the training data
because it would require us to save these subsamples (taking up more space)
despite the fact that the training data is already a random sample (see
\code{\link{learn_cvpairs}}).
}
