% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{learn_cvpairs}
\alias{learn_cvpairs}
\title{Cross-validation pairs}
\usage{
learn_cvpairs(pl, k = NULL, n = 1, test = 0.02)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{k}{Number of folds (an integer).}

\item{n}{Number of test-training pairs to generate (an integer).}

\item{test}{Proportion of observations that should be held out for testing
(a double).}
}
\description{
Add cross validation pairs to a pipelearner object. \code{n} random
partitions will try to be created unless \code{k} is not NULL, in which case
k folds will be created. See \code{\link[modelr]{crossv_mc}} for more
details.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}

