% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_models.R
\name{learn_models}
\alias{learn_models}
\title{Machine learning models}
\usage{
learn_models(pl, models, formulas, ...)
}
\arguments{
\item{pl}{pipelearner object. See \code{\link{pipelearner}}}

\item{models}{Vector of learning model functions which take data and formula
arguments}

\item{formulas}{Vector of objects of class "formula" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{...}{Additional named vectors to be passed to the model function as a
grid of hyperparameters}
}
\description{
Add a learning model to a pipelearner object.
}
\details{
Will expect a pipelearner object, but will also accept a data frame for the
parameter \code{pl}. In the case that a data frame is used, it will first
coerce it to a pipelearner object via \code{\link{pipelearner}}.
}
