% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelearner.R
\name{pipelearner}
\alias{pipelearner}
\title{Create a new pipe learner}
\usage{
pipelearner(data, models = NULL, formulas = NULL, ...)
}
\arguments{
\item{data}{Data frame containing all variables needed by learning algorithm
formulae.}

\item{models}{Vector of learning model functions which take data and formula
arguments}

\item{formulas}{Vector of objects of class "formula" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{...}{Additional named vectors to be passed to the model function as a
grid of hyperparameters}
}
\description{
\code{pipelearner} initializes a new pipelearner object. It is used to
specify the input data frame to subsequent machine learning algorithms.
}
